feat: Added registerPacketIgnored to packet provider
Two new methods were added to the packet provider which now allows the
user to register packets and has no need to be careful with the
PacketAlreadyRegisteredException which prevents clashes between several
packet ids. However the user should be careful because now there is no
longer a warning if some packets are clashing.